+++ src/app/components/signup/signup.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from "../../services/auth.service";
 import { Router } from '@angular/router'
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-signup',
@@ -17,29 +18,40 @@ export class SignupComponent implements OnInit {
   }
 
   error = {
-    err:''
+    err: ''
   }
 
   constructor(private authService: AuthService, private router: Router) { }
 
   ngOnInit(): void {
-    if(this.authService.loggedIn()){
+    if (this.authService.loggedIn()) {
       this.router.navigate(['/home']);
     }
   }
 
-  signUp(){
+  signUp() {
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'center',
+      showConfirmButton: false,
+      timer: 3000,
+      timerProgressBar: true,
+    })
+
     this.authService.signUp(this.user).subscribe(
       res => {
-        if (res.error){
+        if (res.error) {
           this.error.err = res.error
         } else {
-          localStorage.setItem('token', res.token);
-          this.router.navigate(['/profile']);
+          Toast.fire({
+            icon: 'success',
+            title: 'Registro satisfactorio',
+            text: 'Está siendo redireccionado, espere un momento.'
+          }).then(() => {
+            localStorage.setItem('token', res.token);
+            this.router.navigate(['/profile']);
+          })
         }
-      },
-      err => {
-        console.log(err);
       }
     );
   }
